requirements.txt
python==3.9.1
django==3.1.4
djangorestframework==3.12.2
certifi==2020.12.5
wincertstore==0.2
Pillow==8.0.1
django-bootstrap4==2.3.1
